+++ src/helpers/walk-tree.ts
@@ -1,6 +1,6 @@
 import fs from 'fs';
 import paths from 'path';
-import { Ignore } from 'ignore';
+import ignore, { Ignore } from 'ignore';
 
 let allFiles: Set<string>;
 let allFolders: Set<string>;
@@ -14,9 +14,10 @@ interface WalkInput {
 	folderRoots: string[],
 	/** The absolute path of folders being checked */
 	folders: string[],
-	gitignores: Ignore,
-	regexIgnores: RegExp[],
+	/** An instantiated Ignore object listing ignored files */
+	ignored: Ignore,
 };
+
 interface WalkOutput {
 	files: string[],
 	folders: string[],
@@ -24,7 +25,7 @@ interface WalkOutput {
 
 /** Generate list of files in a directory. */
 export default function walk(data: WalkInput): WalkOutput {
-	const { init, commonRoot, folderRoots, folders, gitignores, regexIgnores } = data;
+	const { init, commonRoot, folderRoots, folders, ignored } = data;
 
 	// Initialise files and folders lists
 	if (init) {
@@ -49,18 +50,21 @@ export default function walk(data: WalkInput): WalkOutput {
 			// Create absolute path for disc operations
 			const path = paths.resolve(commonRoot, file).replace(/\\/g, '/');
 			const localPath = localRoot ? file.replace(`./${localRoot}/`, '') : file.replace('./', '');
-			// Skip if nonexistant or ignored
+
+			// Skip if nonexistant
 			const nonExistant = !fs.existsSync(path);
-			const isGitIgnored = gitignores.test(localPath).ignored;
-			const isRegexIgnored = regexIgnores.find(match => localPath.match(match));
-			if (nonExistant || isGitIgnored || isRegexIgnored) continue;
+			if (nonExistant) continue;
+			// Skip if marked as ignored
+			const isIgnored = ignored.test(localPath).ignored;
+			if (isIgnored) continue;
+
 			// Add absolute folder path to list
 			allFolders.add(paths.resolve(folder).replace(/\\/g, '/'));
 			// Check if this is a folder or file
 			if (file.endsWith('/')) {
 				// Recurse into subfolders
 				allFolders.add(path);
-				walk({ init: false, commonRoot: commonRoot, folderRoots, folders: [path], gitignores, regexIgnores });
+				walk({ init: false, commonRoot, folderRoots, folders: [path], ignored });
 			}
 			else {
 				// Add file path to list
@@ -71,7 +75,7 @@ export default function walk(data: WalkInput): WalkOutput {
 	// Recurse into all folders
 	else {
 		for (const i in folders) {
-			walk({ init: false, commonRoot: commonRoot, folderRoots: [folderRoots[i]], folders: [folders[i]], gitignores, regexIgnores });
+			walk({ init: false, commonRoot, folderRoots: [folderRoots[i]], folders: [folders[i]], ignored });
 		}
 	}
 	// Return absolute files and folders lists